+++ static_programs.h
@@ -36,7 +36,7 @@ struct colorwheel_params_t
 {
     uint8_t fade_step;
     uint8_t fade_delay;
-    uint8_t fade_sleep;
+    uint16_t fade_sleep;
     uint16_t hue_start;
     int16_t hue_step;
     uint8_t saturation;
@@ -49,7 +49,7 @@ struct random_params_t
     uint8_t use_address;
     uint8_t fade_step;
     uint8_t fade_delay;
-    uint8_t fade_sleep;
+    uint16_t fade_sleep;
     uint8_t saturation;
     uint8_t value;
     uint8_t min_distance;
